[lib] Implement resyncing of characters
Adds a new module GoonAuth.EVE.Sync that contains a GenServer which can
synchronise users with the CREST API regularly.

The process periodically retrieves all users from LDAP, split between
active and inactive users.

For every active user it checks whether the user's corporation in LDAP
matches the user's corporation according to CREST - if it doesn't the
user's account will be disabled in LDAP and removed from the corporation
group.

In practice, the user won't be able to log in to external services and
will only be able to see the requests subforum.

For every inactive user it checks whether the user's corporation
according to CREST exists in LDAP (i.e. is eligible for auth), and if so
activates the user's account and adds him to the correct group in LDAP.

There are two edge cases to be aware of right now.

1. If a user moves from one authorised corporation to another in the
game (not an option at the moment as we only have Ten Dollar Bond and
alt-corps aren't eligible for auth), GoonAuth will suspend the user's
account in the first cycle and activate it again later on. Effectively
this can mean up to 30 minutes of "downtime" for that user.

2. If a user somehow manages to end up in two corporations in LDAP, or
have an inactive account that is still in a corporation group, GoonAuth
will not currently correct this status. I'll put a Github ticket in.