[lib/jabber] Add Jabber module for pinging
Adds a module using Ejabberd XMLRPC calls to retrieve connected users from an
Ejabberd server and send messages to them.

The message sender is hardcoded as 'trumpbot@tendollarbond.com'

A convenience function exists to message all logged in users automatically, but
it should be easy enough to for example pass a list of LDAP users instead.