[web/auth] Turn authentication & authorization pipeline into a Plug
This change makes it easy to protect additional controllers with authentication and authorization,
simply by inserting plugs at the top.

For example, in order to only let logged-in users access a certain controller and additionally
restrict it to users in an LDAP group called "centipedes", one could add this to the controller:

plug :authenticate
plug :authorize, group: "centipedes"

That's it!