[web] Stateless registrations
Removes the need for an ETS table to keep registration information.

Instead of only retrieving the character once and storing it in ETS, we store the SSO token on the user's session for the duration of the registration. This means one additional call to CREST but no state in GoonAuth, so an overall win.

Removes the cleaner process that took care of removing expired registrations as that is no longer necessary. Registrations expire automatically if the person waits too long with their token.